提交代码

+++ src/main/java/com/cgs/vo/StockValueStaticsVO.java
@@ -9,6 +9,6 @@ import java.util.List;
 public class StockValueStaticsVO {
 
     String date;
-    //市值区分 50亿 100亿 200亿 500亿 1000亿 2000亿 5000亿 1w亿
-    List<Pair<Integer,Double>> valuePairList;
+    //市值区分 10亿 50亿 100亿 200亿 500亿 1000亿 2000亿 5000亿 1w亿
+    List<Pair<Integer,Integer>> valuePairList;
 }
